% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{deaths}
\alias{deaths}
\title{Municipal deaths in Brazil}
\format{
A data frame with 558 rows and 3 variables:
\describe{
  \item{cod}{IBGE microregion code.}
  \item{micro}{Microregion name.}
  \item{ndeaths}{Registered deaths.}
}
}
\source{
DATASUS.
}
\usage{
deaths
}
\description{
Example municipal mortality data.
}
\keyword{datasets}
