% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dtb_info.R
\name{get_dtb}
\alias{get_dtb}
\title{Query the Brazilian Territorial Division}
\usage{
get_dtb(code = NULL, name = NULL)
}
\arguments{
\item{code}{Optional vector of IBGE territorial codes.}

\item{name}{Optional vector of territory names. Matching ignores case and
surrounding whitespace.}
}
\value{
A data frame. If both arguments are `NULL`, the complete table is
  returned. If both are supplied, rows matching either condition are
  returned.
}
\description{
Looks up territorial identifiers and their hierarchy in the version of the
Brazilian Territorial Division (DTB) shipped with the package.
}
\examples{
get_dtb(code = c(50, 5002704, 1))
get_dtb(name = c("Campo Grande", "Recife"))
}
\references{
\url{https://www.ibge.gov.br/geociencias/organizacao-do-territorio/estrutura-territorial/23701-divisao-territorial-brasileira.html}
}
\seealso{
[get_dtb_levels()]
}
