% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{join_brmap}
\alias{join_brmap}
\alias{join_data}
\title{Join tabular data to a Brazilian map}
\usage{
join_brmap(map, data, by = NULL)

join_data(map, data, by = NULL)
}
\arguments{
\item{map}{An `sf`, `SpatialPolygonsDataFrame` or `data.frame` object.}

\item{data}{A data frame containing the columns to add.}

\item{by}{Join specification accepted by [dplyr::left_join()].}
}
\value{
The same broad object type supplied in `map`.
}
\description{
A small type-preserving wrapper around [dplyr::left_join()].
}
\examples{
data("pop2017")
municipalities <- get_brmap("municipality", year = 2023)
municipalities <- join_brmap(
  municipalities, pop2017, by = c("municipality_code" = "mun")
)
}
\seealso{
[get_brmap()]
}
