% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brmap.R
\name{plot_brmap}
\alias{plot_brmap}
\title{Plot a Brazilian map}
\usage{
plot_brmap(
  map,
  data = NULL,
  by = NULL,
  fill_by = NULL,
  theme = theme_brmap(),
  border_colour = "grey30",
  border_linewidth = 0.15,
  fill = "white",
  data_to_join = NULL,
  join_by = NULL,
  var = NULL
)
}
\arguments{
\item{map}{An `sf` or legacy `Spatial` polygon object.}

\item{data}{Optional data frame to join before plotting.}

\item{by}{Join specification accepted by [dplyr::left_join()].}

\item{fill_by}{Optional single column name mapped to polygon fill.}

\item{theme}{A complete or partial ggplot2 theme.}

\item{border_colour}{Polygon border colour.}

\item{border_linewidth}{Polygon border width.}

\item{fill}{Constant fill used when `fill_by` is `NULL`.}

\item{data_to_join, join_by, var}{Deprecated aliases for `data`, `by` and
`fill_by`.}
}
\value{
A `ggplot` object.
}
\description{
Creates a `ggplot` using [ggplot2::geom_sf()]. The returned object can be
extended with any regular ggplot2 layer, scale or label.
}
\examples{
data("pop2017")
south <- get_brmap(
  "municipality", year = 2023, filters = list(region = 4)
)
plot_brmap(
  south,
  data = pop2017,
  by = c("municipality_code" = "mun"),
  fill_by = "pop2017"
)
}
\seealso{
[get_brmap()], [join_brmap()]
}
