% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pop2017}
\alias{pop2017}
\title{Municipal population estimates in 2017}
\format{
A data frame with 5,570 rows and 3 variables:
\describe{
  \item{mun}{IBGE municipality code.}
  \item{nome_mun}{Municipality name.}
  \item{pop2017}{Estimated 2017 population.}
}
}
\source{
Brazilian Institute of Geography and Statistics (IBGE).
}
\usage{
pop2017
}
\description{
Example data containing 2017 municipal population estimates.
}
\keyword{datasets}
