% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{filter.sf}
\alias{reexports}
\alias{plot_sf}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{sf}{\code{\link[sf]{filter.sf}}, \code{\link[sf]{plot_sf}}}
}}

