% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brmap.R
\name{theme_brmap}
\alias{theme_brmap}
\alias{theme_map}
\title{Minimal theme for Brazilian maps}
\usage{
theme_brmap(base_size = 9, base_family = "")

theme_map(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme.
}
\description{
Minimal theme for Brazilian maps
}
