% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gini2015}
\alias{gini2015}
\title{Municipal Gini index in 2015}
\format{
A data frame with 27 rows and 3 variables:
\describe{
  \item{cod}{IBGE state code.}
  \item{uf}{State name.}
  \item{gini}{Gini index.}
}
}
\source{
Brazilian Institute of Geography and Statistics (IBGE).
}
\usage{
gini2015
}
\description{
Example data containing the 2015 municipal Gini index.
}
\keyword{datasets}
