% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gini2015}
\alias{gini2015}
\title{Gini index (2015), state level}
\format{A data frame with 27 rows and 3 variables.}
\usage{
data(gini2015)
}
\description{
IBGE's Gini index of the monthly income distribution of persons 15 years of age
  or over, with income for 2015. \cr\cr
  The data is formatted for easy merging with output from \code{\link[brazilmaps]{get_brmap}}.
}
\details{
\itemize{
  \item \code{cod} The 2-digit code corresponding to the state.
  \item \code{uf} The state name.
  \item \code{gini} The 2015 Gini Index
}
}
\references{
\itemize{
    \item \url{http://www.ibge.gov.br/home/estatistica/pesquisas/pesquisa_resultados.php?id_pesquisa=40}
  }
}
\keyword{data}
